Add files via upload
A XGBoost and a neural network model are used for detecting exoplanets in the deep space. You can visit the link below to download the data: https://www.kaggle.com/keplersmachines/kepler-labelled-time-series-data.

The data are time-series and consist of measurements of the change in flux (light intensity) of thousands of stars gathered from NASA's Kepler space telescope at several time steps.

In this folder you can find pre-trained models of both approaches developed and the corresponding notebook.